add debug versions of plugin+config for prev emu
Add DEBUG versions of the plugins for the previous emulator with
modified code for finding region boundaries (replacing in-line
computation which mis-calculated region boundaries in negative eta
cards), verbose print-out statements, and a modified output which writes
the raw isolation sum IN PLACE of the HoE calculation.

Note that the cmsRun config file also uses an old sample which should be
updated.